fix expand icon margin
