+++ src/components/table/expandable-table/row-expander/row-expander.module.scss
@@ -10,6 +10,7 @@
     display: flex;
     align-items: center;
     color: map-get($color-palette, blue100);
+    margin-right: 5px;
 
     &--expanded {
       transform: rotate(90deg);
