--- src/components/table/expanded-row-content/expanded-row-content.module.scss
@@ -1,6 +1,0 @@
-.expanded-row-content {
-  cursor: pointer;
-  // :first-child {
-  //   padding-left: 42px;
-  // }
-}
