--- src/components/table/expanded-row-content/index.ts
@@ -1,1 +1,0 @@
-export { ExpandedRowContent } from './expanded-row-content';
